Update turn.rb
This took a day to work out due to numerous errors. You can see the #'ed out counter that I tried to add in order to have a loop to enable the code to keep going. I spent a long time trying to get the counter to work, however I failed. I moved on to simply trying to get the code to work for 1 turn and this is what I ended up with. I will implement a loop for turn taking in future lessons.